adds controller code that reacts to pipelinerun status changes
this commit watches for PipelineRun changes and checks if the status is
done. if it is, then it will look up the related SecretAuth object and
delete any objects that were created by that. this allows us to automate
cleaning up of secret auth specific resources when a pipelinerun is
complete.