Env bash and sed arguments
MacOS ships with bash 3 at /bin/bash and rootless security blocks
modifying this binary without shenanigans. Homebrew can install bash 5
but it is in /usr/local/bin/bash . And it's a nice practice to use
/usr/bin/env in shebangs anyway.

BSD sed's -i argument takes a non-optional "extension" to postfix the
file during in-place modification before replacing the original file.
This value should work for both BSD and GNU.

It's also nice practice to add -e for the expression.