Workflow Controller: removes SecretAuth CRD and adds OwnerReferences
The SecretAuth CRD has become redundant with the introduction of
WorkflowRun. The only thing that needed to happen was for the fields
WorkflowRunID and WorkflowID to be moved to WorkflowRun. This commit
adds two new fields to that CRD type: WorkflowRun.Spec.ID and
WorkflowRun.Spec.Workflow.ID. The workflow controller will no longer
watch SecretAuth resources and will just create resources needed for
secret access based on data in the WorkflowRun resource.

This commit also adds the WorkflowRun object as the parent
OwnerReference for every object that gets created for a run. This allows
us to just let Kubernetes garbage collect everything automatically when
the WorkflowRun object is deleted.