Defer invocation resolution to the metadata API
Right now it's easy for a bad function call that only accepts a
parameter to put a run into a completely useless stuck state. This is an
artifact of the lack of error propagation on WorkflowRun CRDs mostly,
but as a stop-gap we can defer the resolution to step execution. This
will at least allow users to see the errors.