refactors step metadata manager
This commit ditches the configmap step metadata manager and uses the
memory implementation. It passes in the metadata via the
KubernetesIntermediaryMetadata type that is configured by the
controller. This type is termporarily available to the metadata api
after vault authenticates the request and builds the scoped manager
factory for the task http request lifecycle.